+++ src/libaktualizr/package_manager/debianmanager_test.cc
@@ -22,6 +22,7 @@ TEST(PackageManagerFactory, Debian_Install_Good) {
   Json::Value target_json;
   target_json["hashes"]["sha256"] = "hash";
   target_json["length"] = 2;
+  target_json["custom"]["ecuIdentifiers"]["primary_serial"]["hardwareId"] = "primary_hwid";
   Uptane::Target target("good.deb", target_json);
 
   storage->storeEcuSerials({{Uptane::EcuSerial("primary_serial"), Uptane::HardwareIdentifier("primary_hwid")}});
@@ -36,7 +37,7 @@ TEST(PackageManagerFactory, Debian_Install_Good) {
   fhandle->wcommit();
 
   EXPECT_EQ(pacman->install(target).result_code.num_code, data::ResultCode::Numeric::kOk);
-  EXPECT_EQ(pacman->getCurrent(), target);
+  EXPECT_TRUE(pacman->getCurrent().MatchTarget(target));
 }
 
 TEST(PackageManagerFactory, Debian_Install_Bad) {
@@ -51,6 +52,7 @@ TEST(PackageManagerFactory, Debian_Install_Bad) {
   Json::Value target_json;
   target_json["hashes"]["sha256"] = "hash";
   target_json["length"] = 2;
+  target_json["custom"]["ecuIdentifiers"]["primary_serial"]["hardwareId"] = "primary_hwid";
   Uptane::Target target("bad.deb", target_json);
 
   std::unique_ptr<StorageTargetWHandle> fhandle = storage->allocateTargetFile(false, target);