+++ src/libaktualizr/uptane/exceptions.h
@@ -71,11 +71,10 @@ class InvalidMetadata : public Exception {
   ~InvalidMetadata() noexcept override = default;
 };
 
-// Currently unused.
 class TargetMismatch : public Exception {
  public:
-  explicit TargetMismatch(const std::string& reponame)
-      : Exception(reponame, "The target metadata in image and director do not match.") {}
+  explicit TargetMismatch(const std::string& targetname)
+      : Exception(targetname, "The target metadata in image and director do not match.") {}
   ~TargetMismatch() noexcept override = default;
 };