+++ tests/uptane_vector_tests.cc
@@ -63,6 +63,8 @@ class VectorWrapper {
     } else if (!vector_["image_repo"]["update"]["is_success"].asBool()) {
       std::cout << "exception from image_repo: '" << vector_["image_repo"]["update"]["err"]
                 << " with message: " << vector_["image_repo"]["update"]["err_msg"] << "\n";
+    } else {
+      std::cout << "an exception while fetching targets metadata.\n";
     }
   }
 
@@ -99,7 +101,8 @@ TEST_P(UptaneVector, Test) {
   Uptane::EcuSerial ecu_serial(config.provision.primary_ecu_serial);
   Uptane::HardwareIdentifier hw_id(config.provision.primary_ecu_hardware_id);
   uptane_client->hw_ids.insert(std::make_pair(ecu_serial, hw_id));
-  Uptane::Target target("test_filename", {{Uptane::Hash::Type::kSha256, "sha256"}}, 1, "");
+  Uptane::EcuMap ecu_map{{ecu_serial, hw_id}};
+  Uptane::Target target("test_filename", ecu_map, {{Uptane::Hash::Type::kSha256, "sha256"}}, 1, "");
   storage->saveInstalledVersion(ecu_serial.ToString(), target, InstalledVersionUpdateMode::kCurrent);
 
   HttpClient http_client;